+++ CHANGELOG.md
@@ -78,6 +78,9 @@ Changes that are not related to specific components
 #### Fixed
 
 - [Foundation] Alert-dark color changed from #d18200->#c27900 
+- Code example syntax highlighting colors to meet WCAG AA contrast requirements
+
+- [Component] What bugs/typos are fixed?
 
 ### Figma